pyproject.toml
[project]
name = "youtube-unsubscriber"
version = "1.0.0"
description = "A script to programmatically manage YouTube subscriptions."
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=0.4.1",
    "psycopg2-binary>=2.9.3",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
youtube-unsubscriber = "youtube_unsubscriber.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
